pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-website-visitor"
dynamic = []
version = "3.0.0"
description = "Automated website visitor with scheduling and advanced browser automation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "nayandas69", email = "nayanchandradas@hotmail.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Monitoring",
]
requires-python = ">=3.10"
dependencies = [
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "schedule>=1.2.0",
    "croniter>=1.4.0",
    "colorama>=0.4.6",
    "requests>=2.31.0",
    "psutil>=5.9.0",
    "packaging>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
awv = "auto_website_visitor.cli:main"
auto-website-visitor = "auto_website_visitor.cli:main"

[project.urls]
Homepage = "https://github.com/nayandas69/auto-website-visitor"
Repository = "https://github.com/nayandas69/auto-website-visitor"
Issues = "https://github.com/nayandas69/auto-website-visitor/issues"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "selenium.*",
    "webdriver_manager.*",
    "schedule.*",
    "croniter.*",
    "colorama.*",
]
ignore_missing_imports = true
